重构Trainer和Tagger，让二者共享代码并支持性能评估
