+++ src/main/java/com/hankcs/hanlp/model/perceptron/PerceptionNERecognizer.java
@@ -11,6 +11,8 @@
 package com.hankcs.hanlp.model.perceptron;
 
 import com.hankcs.hanlp.HanLP;
+import com.hankcs.hanlp.model.perceptron.feature.FeatureMap;
+import com.hankcs.hanlp.model.perceptron.instance.Instance;
 import com.hankcs.hanlp.model.perceptron.model.LinearModel;
 import com.hankcs.hanlp.model.perceptron.tagset.NERTagSet;
 import com.hankcs.hanlp.model.perceptron.common.TaskType;
@@ -45,6 +47,7 @@ public class PerceptionNERecognizer extends PerceptronTagger
 
     /**
      * 加载配置文件指定的模型
+     *
      * @throws IOException
      */
     public PerceptionNERecognizer() throws IOException
@@ -73,8 +76,8 @@ public class PerceptionNERecognizer extends PerceptronTagger
     }
 
     @Override
-    public boolean learn(Sentence sentence)
+    protected Instance createInstance(Sentence sentence, FeatureMap featureMap)
     {
-        return learn(NERInstance.create(sentence, model.featureMap));
+        return NERInstance.create(sentence, featureMap);
     }
 }
