+++ src/main/java/com/hankcs/hanlp/model/perceptron/PerceptronPOSTagger.java
@@ -11,6 +11,8 @@
 package com.hankcs.hanlp.model.perceptron;
 
 import com.hankcs.hanlp.HanLP;
+import com.hankcs.hanlp.model.perceptron.feature.FeatureMap;
+import com.hankcs.hanlp.model.perceptron.instance.Instance;
 import com.hankcs.hanlp.model.perceptron.instance.POSInstance;
 import com.hankcs.hanlp.model.perceptron.model.LinearModel;
 import com.hankcs.hanlp.model.perceptron.common.TaskType;
@@ -108,8 +110,8 @@ public class PerceptronPOSTagger extends PerceptronTagger
     }
 
     @Override
-    public boolean learn(Sentence sentence)
+    protected Instance createInstance(Sentence sentence, FeatureMap featureMap)
     {
-        return learn(POSInstance.create(sentence, model.featureMap));
+        return POSInstance.create(sentence, featureMap);
     }
 }
