+++ src/main/java/com/hankcs/hanlp/model/perceptron/PerceptronSegmenter.java
@@ -12,6 +12,7 @@
 package com.hankcs.hanlp.model.perceptron;
 
 import com.hankcs.hanlp.HanLP;
+import com.hankcs.hanlp.model.perceptron.feature.FeatureMap;
 import com.hankcs.hanlp.model.perceptron.instance.CWSInstance;
 import com.hankcs.hanlp.model.perceptron.model.LinearModel;
 import com.hankcs.hanlp.model.perceptron.common.TaskType;
@@ -125,8 +126,16 @@ public class PerceptronSegmenter extends PerceptronTagger
     }
 
     @Override
-    public boolean learn(Sentence sentence)
+    protected Instance createInstance(Sentence sentence, FeatureMap featureMap)
     {
-        return learn(CWSInstance.create(sentence, model.featureMap));
+        return CWSInstance.create(sentence, featureMap);
+    }
+
+    @Override
+    public double[] evaluate(String corpora) throws IOException
+    {
+        // 这里用CWS的F1
+        double[] prf = Utility.prf(Utility.evaluateCWS(corpora, this));
+        return prf;
     }
 }
