+++ src/main/java/com/hankcs/hanlp/model/perceptron/utility/Utility.java
@@ -11,6 +11,7 @@
  */
 package com.hankcs.hanlp.model.perceptron.utility;
 
+import com.hankcs.hanlp.model.perceptron.PerceptronSegmenter;
 import com.hankcs.hanlp.model.perceptron.instance.Instance;
 import com.hankcs.hanlp.corpus.document.CorpusLoader;
 import com.hankcs.hanlp.corpus.document.Document;
@@ -18,15 +19,13 @@ import com.hankcs.hanlp.corpus.document.sentence.Sentence;
 import com.hankcs.hanlp.corpus.document.sentence.word.CompoundWord;
 import com.hankcs.hanlp.corpus.document.sentence.word.IWord;
 import com.hankcs.hanlp.corpus.document.sentence.word.Word;
+import com.hankcs.hanlp.model.perceptron.instance.InstanceHandler;
 
 import java.io.BufferedWriter;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStreamWriter;
-import java.util.ArrayList;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Random;
+import java.util.*;
 
 /**
  * @author hankcs
@@ -268,5 +267,63 @@ public class Utility
         }
 
         return wordArray;
+    }
+
+    public static int[] evaluateCWS(String developFile, final PerceptronSegmenter segmenter) throws IOException
+    {
+        // int goldTotal = 0, predTotal = 0, correct = 0;
+        final int[] stat = new int[3];
+        Arrays.fill(stat, 0);
+        IOUtility.loadInstance(developFile, new InstanceHandler()
+        {
+            @Override
+            public boolean process(Sentence sentence)
+            {
+                List<Word> wordList = toSimpleWordList(sentence);
+                String[] wordArray = toWordArray(wordList);
+                stat[0] += wordArray.length;
+                String text = com.hankcs.hanlp.utility.TextUtility.combine(wordArray);
+                String[] predArray = segmenter.segment(text).toArray(new String[0]);
+                stat[1] += predArray.length;
+
+                int goldIndex = 0, predIndex = 0;
+                int goldLen = 0, predLen = 0;
+
+                while (goldIndex < wordArray.length && predIndex < predArray.length)
+                {
+                    if (goldLen == predLen)
+                    {
+                        if (wordArray[goldIndex].equals(predArray[predIndex]))
+                        {
+                            stat[2]++;
+                            goldLen += wordArray[goldIndex].length();
+                            predLen += wordArray[goldIndex].length();
+                            goldIndex++;
+                            predIndex++;
+                        }
+                        else
+                        {
+                            goldLen += wordArray[goldIndex].length();
+                            predLen += predArray[predIndex].length();
+                            goldIndex++;
+                            predIndex++;
+                        }
+                    }
+                    else if (goldLen < predLen)
+                    {
+                        goldLen += wordArray[goldIndex].length();
+                        goldIndex++;
+                    }
+                    else
+                    {
+                        predLen += predArray[predIndex].length();
+                        predIndex++;
+                    }
+                }
+
+                return false;
+            }
+        });
+        return stat;
     }
 }
